% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilrot.R
\name{bilrot}
\alias{bilrot}
\title{bilrot takes an image and rotates it by theta using bilinear interpolation to fill
in the gaps.}
\usage{
bilrot(theta, image)
}
\arguments{
\item{theta}{is the angle by which you want to rotate (in degrees)}

\item{image}{is the 2D array of depth values}
}
\value{
a 2d matrix with the updated depth values
}
\description{
bilrot takes an image and rotates it by theta using bilinear interpolation to fill
in the gaps.
}
\examples{
rotImage <- bilrot(45, matrix(c(1,2,3,4), nrow = 2))

}

