% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeImage.R
\name{completeImage}
\alias{completeImage}
\title{completeImage takes as input a 2D array of depth values which may contain NA
values, and applies completeSeq row wise, and then column wise, and then takes
the average of the two resulting outputs.}
\usage{
completeImage(image)
}
\arguments{
\item{image}{is the 2D array of depth values}
}
\value{
image with all the NA values replaced by linearly interpolated values
in the y direction averaged with the linearly interpolated values in the x direction
}
\description{
completeImage takes as input a 2D array of depth values which may contain NA
values, and applies completeSeq row wise, and then column wise, and then takes
the average of the two resulting outputs.
}
\examples{
completedIm <- completeImage(matrix(c(1,2,3,4), nrow = 2))

}

