% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeSeq.R
\name{completeSeq}
\alias{completeSeq}
\title{completeSeq takes a vector of numbers that may contain NA values and recursively
fills in the NA values by interpolating the about a symmetric string of NA's. This
method assumes the first and last element of the vector to be non NA values.}
\usage{
completeSeq(row)
}
\arguments{
\item{row}{is the vector whose first and last elements are non NA values}
}
\value{
row but all the NA's are linearly interpolated
}
\description{
completeSeq takes a vector of numbers that may contain NA values and recursively
fills in the NA values by interpolating the about a symmetric string of NA's. This
method assumes the first and last element of the vector to be non NA values.
}
\examples{
completedVec <- completeSeq(c(1,NA,NA,3,NA,NA,NA,9))

}

