% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewDepths.R
\name{getNewDepths}
\alias{getNewDepths}
\title{getNewDepths gives a 2d array of adjusted depth values}
\usage{
getNewDepths(image, newDom, plane, P)
}
\arguments{
\item{image}{is the original 2d matrix with depth values representing the 3d image}

\item{newDom}{is a data.frame where the first column ($newCoordi) are the new i coords,
and the second column ($newCoordj) contains the new j coords}

\item{plane}{is data.frame containing a b and c where z = a + bx + cy}

\item{P}{is the 2d matrix containing shortest distances from each point to plane}
}
\value{
a 2d matrix with the updated depth values
}
\description{
getNewDepths gives a 2d array of adjusted depth values
}

