% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewDepthsApprox.R
\name{getNewDepthsApprox}
\alias{getNewDepthsApprox}
\title{getNewDepthsApprox gives a 2d array of adjusted depth values where the new best
fit plane is a level plane}
\usage{
getNewDepthsApprox(image, plane)
}
\arguments{
\item{image}{is the original 2d matrix with depth values representing the 3d image}

\item{plane}{is data.frame containing a b and c where z = a + bx + cy}
}
\value{
a 2d matrix with the updated depth values
}
\description{
getNewDepthsApprox gives a 2d array of adjusted depth values where the new best
fit plane is a level plane
}
\examples{
getNewDepthsApprox <-
getNewDepthsApprox(matrix(c(1,2,3,4), nrow = 2),
data.frame(a = 1, b = 2, c = 3))

}

