% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewDomain.R
\name{getNewDomain}
\alias{getNewDomain}
\title{getNewDomain gives a vector where the ith row has the new i and j coordinates}
\usage{
getNewDomain(image, plane)
}
\arguments{
\item{image}{is a 2d matrix with depth values representing the 3d image}

\item{plane}{is data.frame containing a b and c where z = a + bx + cy}
}
\value{
a data.frame with two columns, the first being a mapping from 1:size(image) to the
new i coordinate domain, the second being a mapping from 1:size(image) to the new j
coordinate domain

(ignore above return) a 2d matrix where the ijth entry is the new ij coordinate
}
\description{
getNewDomain gives a vector where the ith row has the new i and j coordinates
}
\examples{
newDom <- getNewDomain(matrix(c(1,2,3,4), nrow = 2),
data.frame(a = 141.8, b = .0584, c = -0.0146))

}

