% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlane.R
\name{getPlane}
\alias{getPlane}
\title{getPlane gets the fitted coefficients for the entire image}
\usage{
getPlane(image)
}
\arguments{
\item{image}{is a 2d matrix with depth values representing the 3d image}
}
\value{
A data frame with 3 elements. The "a", "b", and "c" components are the parameters of the
plane which takes the form z = a + bx + cy
}
\description{
Do a multiple regression on the depth values of the basis functions and
return the coefficients a, b, and c
}
\examples{
coeffs <- getPlane(matrix(c(1,2,3,4), nrow = 2))

}

