% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShortestDist.R
\name{getShortestDist}
\alias{getShortestDist}
\title{getShortestDist gives a 2d matrix where the ij entry is the shortest distance
from that point and depth to the plane given by the parameters a b and c}
\usage{
getShortestDist(image, plane)
}
\arguments{
\item{image}{is a 2d matrix with depth values representing the 3d image}

\item{plane}{is data.frame containing a b and c where z = a + bx + cy}
}
\value{
a 2d matrix containing the shortest distances
}
\description{
getShortestDist gives a 2d matrix where the ij entry is the shortest distance
from that point and depth to the plane given by the parameters a b and c
}
\examples{
dists <- getShortestDist(matrix(c(1,2,3,4), nrow = 2),
data.frame(a = 141.8, b = .0584, c = -0.0146))

}

