% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeExtreme.R
\name{removeExtreme}
\alias{removeExtreme}
\title{removeExtreme removes extreme parts of an image}
\usage{
removeExtreme(image)
}
\arguments{
\item{image}{is the original 2d array of depth values}
}
\value{
the same image but with the extreme depth values removed
}
\description{
removeExtreme removes extreme parts of an image
}
\examples{
breechFace <- removeExtreme(matrix(c(1,2,3,4), nrow = 2))

}

