% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmFPavg.R
\name{rmFPavg}
\alias{rmFPavg}
\title{rmFPavg removes the firing pin from the 3d image by finding the average pixel depth,
and then removing all pixels with a depth greater than .853 * (avg pixel depth)}
\usage{
rmFPavg(image)
}
\arguments{
\item{image}{is the original 2d array of depth values}
}
\value{
the same image but with the firing pin removed
}
\description{
rmFPavg removes the firing pin from the 3d image by finding the average pixel depth,
and then removing all pixels with a depth greater than .853 * (avg pixel depth)
}
\examples{
breechFace <- rmFPavg(matrix(c(1,2,3,4), nrow = 2))

}

