% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmFPradius.R
\name{rmFPradius}
\alias{rmFPradius}
\title{rmFPradius removes the firing pin from the 3d image by finding the center coordinates,
and then removing all pixels within a specified radius of that center.}
\usage{
rmFPradius(image, rad = 145000)
}
\arguments{
\item{image}{is the original 2d array of depth values}

\item{rad}{is the radius of the circle centered at the firing pin center
you want to filter out}
}
\value{
the same image but with the firing pin removed
}
\description{
rmFPradius removes the firing pin from the 3d image by finding the center coordinates,
and then removing all pixels within a specified radius of that center.
}
\examples{
breechFace <- rmFPradius(matrix(c(1,2,3,4), nrow = 2))

}

