% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{rotate runs the entire rotation process}
\usage{
rotate(x3pFile)
}
\arguments{
\item{x3pFile}{is the original list vector of data whose first element is the
metadata, and the second element is the 2D array of depth values.}
}
\value{
the same file, but the 2D array of depth values is now leveled out.
}
\description{
rotate runs the entire rotation process
}

